Trait slos_filesystem::FsNode

pub trait FsNode: Debug {
    fn mount(&self) -> Option<&dyn FsRoot>;

    fn inode(&self) -> usize;

    fn name(&self) -> &str;

    fn permissions(&self) -> u16;

    fn try_directory(&mut self) -> Option<&mut dyn FsDirectory>;

    fn try_file(&mut self) -> Option<&mut dyn FsFile>;
}

Filesystem node

Required methods

fn mount(&self) -> Option<&dyn FsRoot>

Try to get the root filesystem this node belongs to

fn inode(&self) -> usize

Get the inode value for this node

fn name(&self) -> &str

Get the filename of this node

fn permissions(&self) -> u16

Get the permissions of this node

fn try_directory(&mut self) -> Option<&mut dyn FsDirectory>

Try to get this node as a FsDirectory trait object reference

Will always return None if this node is a file.

fn try_file(&mut self) -> Option<&mut dyn FsFile>

Try to get this node as a FsFile trait object reference

Will always return None if this node is a directory.